+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 
 - Currey and Co. Base template
+- Currey and Co. Shipping information (single product shipping)
+- Currey and Co. Warranty information (single product shipping)
 
 ## [ released ]